Add interactive demo environment with OAuth2 flow demonstrations (#15)
* Add interactive demo environment with OAuth2 flow demonstrations

- Add demo controllers for Auth Code, PKCE, Client Credentials flows
- Add JWT inspector and API playground pages
- Add Telescope integration for debugging/observability
- Add AssignUserRole console command
- Create demo manual documentation with screenshots
- Add demo-specific React components and layouts
- Update existing UI components and layouts for demo support
- Add feature tests for demo functionality

* Add Client Manager to demo section

- Add ClientManagerController for creating/managing OAuth demo clients
- Add ClientManager React component with create/delete functionality
- Add flash message support to Inertia shared data
- Add custom JWT bearer token response
- Link to manual.md from README.md
- Add Client Manager navigation and card to demo index

+++ app/Services/SigningKeyService.php
@@ -29,7 +29,7 @@ class SigningKeyService
         $privateKeyPem = $privateKey->toString('PKCS8');
         $publicKeyPem = $publicKey->toString('PKCS8');
 
-        $lifetimeDays = config('identity-gateway.key_rotation.lifetime_days', 90);
+        $lifetimeDays = (int) config('identity-gateway.key_rotation.lifetime_days', 90);
 
         return SigningKey::create([
             'private_key' => $privateKeyPem,
